+++ src/lib/mcp/prompts.ts
@@ -22,11 +22,8 @@ import { z } from "zod/v4";
 
 import { prisma } from "@/lib/db";
 import { annotate } from "@/lib/logging/context";
-import {
-  collectDoctorReportData,
-  type DoctorReportData,
-} from "@/lib/doctor-report-data";
-import { getMetricStatusMeta } from "@/lib/insights/metric-status-registry";
+import { collectDoctorReportData } from "@/lib/doctor-report-data";
+import { summariseForVisit } from "./doctor-visit-summary";
 import { coachScopeWindowSchema } from "@/lib/ai/coach/types";
 import type { CoachScopeWindow } from "@/lib/ai/coach/types";
 import { executeCoachTool } from "@/lib/ai/coach/tools/executor";
@@ -69,136 +66,6 @@ export interface McpPromptDefinition {
     ctx: McpAuthContext,
     args: Record<string, unknown>,
   ) => Promise<McpPromptResult>;
-}
-
-/**
- * Fallback units for the headline specialised metric types the generic
- * `metric-status-registry` intentionally omits (it carries the synced/additive
- * metrics only). The doctor-report stats map is keyed by `MeasurementType`.
- */
-const FALLBACK_UNITS: Record<string, string> = {
-  WEIGHT: "kg",
-  BLOOD_PRESSURE_SYS: "mmHg",
-  BLOOD_PRESSURE_DIA: "mmHg",
-  PULSE: "bpm",
-  BODY_MASS_INDEX: "kg/m²",
-  SLEEP_DURATION: "min",
-};
-
-function unitAndBandFor(type: string): {
-  unit: string;
-  referenceBand: { low: number; high: number } | null;
-} {
-  const meta = getMetricStatusMeta(type);
-  if (meta) {
-    return {
-      unit: meta.unit,
-      referenceBand: meta.normalRange
-        ? { low: meta.normalRange.low, high: meta.normalRange.high }
-        : null,
-    };
-  }
-  return { unit: FALLBACK_UNITS[type] ?? "", referenceBand: null };
-}
-
-const round = (n: number): number => Math.round(n * 100) / 100;
-
-/**
- * Reduce the full doctor-report payload to a compact, clinician-oriented
- * summary. Latest + aggregate per metric, latest reading per analyte — never
- * the full per-reading history (R-DEL-2). Every numeric value is one the server
- * computed; every band is server-side. No verdict, no diagnosis.
- */
-function summariseForVisit(data: DoctorReportData): Record<string, unknown> {
-  const vitals = Object.entries(data.stats).map(([type, s]) => {
-    const { unit, referenceBand } = unitAndBandFor(type);
-    return {
-      metric: type,
-      unit,
-      referenceBand,
-      latest: round(s.latest),
-      avg: round(s.avg),
-      min: round(s.min),
-      max: round(s.max),
-      readings: s.count,
-    };
-  });
-
-  const compliance = Object.entries(data.compliance).map(([name, c]) => ({
-    medication: name,
-    takenDoses: c.taken,
-    expectedDoses: c.total,
-    missedDoses: c.missed,
-    adherencePct:
-      c.total > 0 ? Math.min(100, Math.round((c.taken / c.total) * 100)) : null,
-  }));
-
-  const medications = data.medications
-    .slice(0, 40)
-    .map((m) => ({ name: m.name, dose: m.dose }));
-
-  const labs = (data.labResults ?? []).slice(0, 50).map((l) => ({
-    analyte: l.analyte,
-    panel: l.panel,
-    value: l.value,
-    valueText: l.valueText,
-    unit: l.unit,
-    referenceLow: l.referenceLow,
-    referenceHigh: l.referenceHigh,
-    takenAt: l.takenAt,
-    readings: l.count,
-  }));
-
-  const wellness = (data.wellnessScores ?? []).map((w) => ({
-    score: w.type,
-    latest: round(w.latest),
-    avg: round(w.avg),
-    note: "descriptive composite, not a clinical assessment",
-  }));
-
-  const illness = (data.illnessEpisodes ?? []).map((e) => ({
-    label: e.label,
-    type: e.type,
-    lifecycle: e.lifecycle,
-    onsetAt: e.onsetAt,
-    resolvedAt: e.resolvedAt,
-  }));
-
-  return {
-    period: {
-      days: data.period.days,
-      start: data.period.start,
-      end: data.period.end,
-    },
-    patient: {
-      dateOfBirth: data.patient.dateOfBirth,
-      sex: data.patient.gender,
-      heightCm: data.patient.heightCm,
-    },
-    bmi: data.bmi,
-    vitals,
-    medications,
-    compliance,
-    // Mood is privacy-gated upstream; include only the aggregate when present.
-    ...(data.mood
-      ? {
-          mood: {
-            avg: round(data.mood.avg),
-            min: data.mood.min,
-            max: data.mood.max,
-            entries: data.mood.count,
-          },
-        }
-      : {}),
-    // Glucose clinical panel only when the user logs glucose (stillLearning
-    // flags a too-thin period — surfaced honestly, never asserted).
-    ...(data.glucoseClinical && !data.glucoseClinical.stillLearning
-      ? { glucosePanel: data.glucoseClinical }
-      : {}),
-    ...(labs.length > 0 ? { labs } : {}),
-    ...(wellness.length > 0 ? { wellnessScores: wellness } : {}),
-    ...(illness.length > 0 ? { illnessEpisodes: illness } : {}),
-  };
 }
 
 /**
